rfc5280: Move duplicated extensions test to rfc5280
