+++ limbo/testcases/__init__.py
@@ -4,7 +4,6 @@ Models and definitions for generating Limbo testcases.
 
 from limbo.testcases._core import registry
 
-from .basic import *  # noqa: F403
 from .pathlen import *  # noqa: F403
 from .rfc5280 import *  # noqa: F403
 from .webpki import *  # noqa: F403
