+++ limbo/testcases/rfc5280.py
@@ -1445,3 +1445,30 @@ def serial_number_zero(builder: Builder) -> None:
     builder.trusted_certs(root).peer_certificate(leaf).expected_peer_name(
         PeerName(kind="DNS", value="example.com")
     ).fails()
+
+
+@testcase
+def duplicate_extensions(builder: Builder) -> None:
+    """
+    Produces the following **invalid** chain:
+
+    ```
+    root -> EE
+    ```
+
+    This chain is invalid solely because of the EE cert's construction:
+    it contains multiple X.509v3 extensions with the same OID, which
+    is prohibited
+    """
+
+    root = builder.root_ca()
+    leaf = builder.leaf_cert(
+        root,
+        san=ext(x509.SubjectAlternativeName([x509.DNSName("example.com")]), critical=False),
+        extra_extension=ext(
+            x509.SubjectAlternativeName([x509.DNSName("example.com")]), critical=False
+        ),
+    )
+
+    builder = builder.server_validation()
+    builder = builder.trusted_certs(root).peer_certificate(leaf).fails()
